feat: 动态判断是否已签名

+++ lib/fastlane/plugin/android_channels/actions/android_channels_action.rb
@@ -7,42 +7,54 @@ module Fastlane
   module Actions
     class AndroidChannelsAction < Action
       def self.run(params)
-        unsigned_apk = Helper::AndroidChannelsHelper.determine_unsigned_apk!(params)
-        keystore = Helper::AndroidChannelsHelper.determine_keystore!(params)
-        output_path = Helper::AndroidChannelsHelper.determine_output_path!(params)
+        apk_file = Helper::AndroidChannelsHelper.determine_apk_file!(params)
         channels = Helper::AndroidChannelsHelper.determine_channels!(params)
-        apksigner = Helper::AndroidChannelsHelper.determine_apksigner!(params)
-        keystore_password = Helper::AndroidChannelsHelper.determine_keystore_password!(params)
-        key_alias = params[:key_alias]
-        key_password = params[:key_password]
+        output_path = Helper::AndroidChannelsHelper.determine_output_path!(params)
 
-        FastlaneCore::PrintTable.print_values(
-          title: "Summary for android_channels #{AndroidChannels::VERSION}",
-          config: {
+        signed = Helper::AndroidChannelsHelper.is_signed?(apk_file)
+        summary_params = {
+          apk_file: apk_file,
+          signed: signed,
+          channels: channels,
+          output_path: output_path,
+        }
+
+        unless signed
+          keystore = Helper::AndroidChannelsHelper.determine_keystore!(params)
+          apksigner = Helper::AndroidChannelsHelper.determine_apksigner!(params)
+          keystore_password = Helper::AndroidChannelsHelper.determine_keystore_password!(params)
+          key_alias = params[:key_alias]
+          key_password = params[:key_password]
+
+          summary_params.merge!({
             apksigner: apksigner,
-            unsigned_apk: unsigned_apk,
-            channels: channels,
             keystore: keystore,
             keystore_password: keystore_password,
             key_alias: key_alias,
-            key_password: key_password,
-            output_path: output_path,
-            clean: params[:clean]
-          },
+            key_password: key_password
+          })
+        end
+
+        FastlaneCore::PrintTable.print_values(
+          title: "Summary for android_channels #{AndroidChannels::VERSION}",
+          config: summary_params,
           mask_keys: %i[keystore_password key_password]
         )
 
-        UI.verbose "Signing apk ..."
-        signed_file = Helper::AndroidChannelsHelper.sign_apk(apksigner, keystore, keystore_password, unsigned_apk, key_alias, key_password)
+        if signed
+          UI.verbose "Packaging channel apk ..."
+          Helper::AndroidChannelsHelper.write_apk_with_channels(output_path, apk_file, channels)
+        else
+          UI.verbose "Signing apk ..."
+          signed_file = Helper::AndroidChannelsHelper.sign_apk(apksigner, keystore, keystore_password, apk_file, key_alias, key_password)
 
-        UI.verbose "Packaging channel apk ..."
-        Helper::AndroidChannelsHelper.write_apk_with_channels(output_path, signed_file.path, channels)
+          UI.verbose "Packaging channel apk ..."
+          Helper::AndroidChannelsHelper.write_apk_with_channels(output_path, signed_file.path, channels)
+          signed_file.unlink
+        end
 
         total = Dir["#{output_path}/*"].size
         UI.success "Packaged done, total: #{total} apk file(s)."
-
-        signed_file.close
-        signed_file.unlink
 
         output_path
       end
@@ -65,21 +77,10 @@ module Fastlane
 
       def self.available_options
         [
-          FastlaneCore::ConfigItem.new(key: :android_sdk_path,
-                                  env_name: "ANDROID_CHANNELS_ANDROID_SDK_PATH",
-                               description: "The path of android sdk",
-                                  optional: false,
-                             default_value: ENV["ANDROID_SDK_ROOT"] || ENV["ANDROID_HOME"],
-                                      type: String),
-          FastlaneCore::ConfigItem.new(key: :build_tools_version,
-                                  env_name: "ANDROID_CHANNELS_BUILD_TOOLS_VERSION",
-                               description: "The version of build tools (by default, use latest one)",
-                                  optional: true,
-                                      type: String),
-          FastlaneCore::ConfigItem.new(key: :unsigned_apk,
-                                  env_name: "ANDROID_CHANNELS_UNSIGNED_APK",
-                               description: "The path of unsigned apk file",
-                             default_value: Dir['**/*'].select{|f|f.end_with?('-unsigned.apk')}.last,
+          FastlaneCore::ConfigItem.new(key: :apk_file,
+                                  env_name: "ANDROID_CHANNELS_APK_FILE",
+                               description: "The path of apk file",
+                             default_value: Dir['**/*'].select{|f|f.end_with?('signed.apk')}.last,
                                   optional: false,
                                       type: String),
           FastlaneCore::ConfigItem.new(key: :channels,
@@ -87,32 +88,43 @@ module Fastlane
                                description: "The key password of keystore",
                                   optional: false,
                                       type: Array),
+          FastlaneCore::ConfigItem.new(key: :output_path,
+                                  env_name: "ANDROID_CHANNELS_OUTPUT_PATH",
+                               description: "The output path of channels apk files",
+                                  optional: true,
+                             default_value: "channels_apk",
+                                      type: String),
+          FastlaneCore::ConfigItem.new(key: :android_sdk_path,
+                                  env_name: "ANDROID_CHANNELS_ANDROID_SDK_PATH",
+                               description: "The path of android sdk",
+                             default_value: ENV["ANDROID_SDK_ROOT"] || ENV["ANDROID_HOME"],
+                                  optional: true,
+                                      type: String),
+          FastlaneCore::ConfigItem.new(key: :build_tools_version,
+                                  env_name: "ANDROID_CHANNELS_BUILD_TOOLS_VERSION",
+                               description: "The version of build tools (by default, always use latest version)",
+                                  optional: true,
+                                      type: String),
           FastlaneCore::ConfigItem.new(key: :keystore,
                                   env_name: "ANDROID_CHANNELS_KEYSTORE",
                                description: "The path of keystore file",
                              default_value: Dir['**/*'].select{|f|f.end_with?('.keystore')}.last,
-                                  optional: false,
+                                  optional: true,
                                       type: String),
           FastlaneCore::ConfigItem.new(key: :keystore_password,
                                   env_name: "ANDROID_CHANNELS_KEYSTORE_PASSWORD",
                                description: "The password of keystore",
-                                  optional: false,
+                                  optional: true,
                                       type: String),
           FastlaneCore::ConfigItem.new(key: :key_alias,
                                   env_name: "ANDROID_CHANNELS_KEY_ALIAS",
                                description: "The key alias of keystore",
-                                  optional: false,
+                                  optional: true,
                                       type: String),
           FastlaneCore::ConfigItem.new(key: :key_password,
                                   env_name: "ANDROID_CHANNELS_KEY_PASSWORD",
                                description: "The key password of keystore",
-                                  optional: false,
-                                      type: String),
-          FastlaneCore::ConfigItem.new(key: :output_path,
-                                  env_name: "ANDROID_CHANNELS_OUTPUT_PATH",
-                               description: "The output of signed files",
                                   optional: true,
-                             default_value: "signed_apk",
                                       type: String),
           FastlaneCore::ConfigItem.new(key: :clean,
                                   env_name: "ANDROID_CHANNELS_CLEAN",
@@ -134,7 +146,14 @@ module Fastlane
       def self.example_code
         [
           'android_channels(
-            unsigned_apk: "app-unsigned.apk",
+            apk_file: "app-signed.apk",
+            channels: ["xiaomi", "huawei", "qq"],
+            output_path: "/tmp/output_path",
+          )',
+          'android_channels(
+            apk_file: "app-unsigned.apk",
+            signed: false,
+            channels: ["xiaomi", "huawei", "qq"],
             keystore: "release.keystore",
             keystore_password: "p@ssword",
             clean: true
